Use shared author image list
