+++ src/components/pages/students/featured-project.js
@@ -3,30 +3,17 @@ import dlv from 'dlv';
 import { useStaticQuery, graphql } from 'gatsby';
 import styled from '@emotion/styled';
 import Link from '~components/dev-hub/link';
-import AuthorImage from '~components/dev-hub/author-image';
+import AuthorImageList from '~components/dev-hub/author-image-list';
 import Badge from '~components/dev-hub/badge';
 import MediaBlock from '~components/dev-hub/media-block';
 import { H5, P } from '~components/dev-hub/text';
 import { size } from '~components/dev-hub/theme';
 import { transformProjectStrapiData } from '~utils/transform-project-strapi-data';
 
-const AUTHOR_IMAGE_HEIGHT = 30;
-
 const RelativePositionedBadge = styled(Badge)`
     margin-left: 0;
     position: relative;
     width: fit-content;
-`;
-
-const AuthorImageContainer = styled('div')`
-    display: flex;
-`;
-
-const StyledAuthorImage = styled(AuthorImage)`
-    margin-right: ${size.small};
-    :not(:last-of-type) {
-        margin-right: -${size.small};
-    }
 `;
 
 const DescriptionText = styled(P)`
@@ -36,21 +23,6 @@ const DescriptionText = styled(P)`
 const FeaturedImage = styled('img')`
     border-radius: ${size.xsmall};
 `;
-
-const AuthorImages = ({ students }) => (
-    <AuthorImageContainer>
-        {students.map(({ name, image_url }) => (
-            <StyledAuthorImage
-                gradientOffset={4}
-                hideOnMobile={false}
-                height={AUTHOR_IMAGE_HEIGHT}
-                width={AUTHOR_IMAGE_HEIGHT}
-                image={image_url}
-                key={name}
-            />
-        ))}
-    </AuthorImageContainer>
-);
 
 const galleryFeaturedProject = graphql`
     query GalleryFeaturedProject {
@@ -106,7 +78,7 @@ const FeaturedProject = () => {
             <RelativePositionedBadge contentType="featured" />
             <H5>{name}</H5>
             <DescriptionText>{description}</DescriptionText>
-            <AuthorImages students={students} />
+            <AuthorImageList size={30} students={students} />
             <Link tertiary to={slug}>
                 View Project
             </Link>
